fix: use usize for SyncResult.reconciled to avoid lossy u64 cast
reconciled was u64 (from sqlx rows_affected) but every consumer cast
it to usize for arithmetic with changed.len(). Store it as usize at
the boundary to eliminate the cast at every call site.

+++ src/api/sync.rs
@@ -38,7 +38,7 @@ pub async fn post_sync(State(state): State<AppState>) -> Result<StatusCode, AppE
                 changed,
                 reconciled,
             }) => {
-                let count = changed.len() + reconciled as usize;
+                let count = changed.len() + reconciled;
                 if count > 0 {
                     let _ = state_clone.tx.send(SyncEvent::NewNotifications { count });
                     auto_fetch_viewport_prs(&state_clone, &state_clone.tx, &changed).await;
